+++ odoo/local-src/specific_report/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from . import invoice
+from . import purchase
+from . import repair
+from . import sale