+++ HISTORY.rst
@@ -20,6 +20,17 @@ Unreleased
 ++++++++++
 
 **Features and Improvements**
+- Add l10m_ch_bank
+- Add l10n_ch_zip
+- Add  l10n_ch_states
+- Port sales_product_set
+- Port sales_product_set_layout
+- Add specific_report:
+  - Quotation / Sale Order.  Note: the order of columns has changed
+  - Invoice.  Note: There was two fields referring to the same place "Reference /  Description". This is now on the field "Description".
+  - Delivery.  Note: Some changes in the headers
+  - Purchase Order.  Note: there is not anymore the "validate by:" field
+  - Mrp / Quotation.  Note: 'Lot id' label is diplayed only if there is one.
 
 **Bugfixes**